GesoTownPresenter内の「前回利用日から日付が変わっていないか」という処理をModelに移行しました
